refactor: split browser-pure lexicon core
